// src/app/enums/enums.ts
export enum chartNumber{
    NAME = 76281,  //1000 Character Names for Fantasy (d1000)
    DUNGEON = 20234,  //Random dungeon name generator
    TAVERN = 114, //Tavern Name Generator (d20)
    SPELL = 46433 //Spell Name (1d2)
}

export enum enemyActions{
    ATTACK,
    STRONG_ATTACK,
    DEFEND,
    STRONG_DEFEND,
    DEBUFF,
    STRONG_DEBUFF,
    BUFF,
    STRONG_BUFF,
}
export enum playerActions{
    ATTACK,
    GUARD,
    HEALTH_UP
}

export enum actionIcons{
    ATTACK = "/assets/game-assets/action-icons/attack.png",
    STRONG_ATTACK = "/assets/game-assets/action-icons/strong_attack.png",
    DEFEND = "/assets/game-assets/action-icons/defend.png",
    STRONG_DEFEND = "/assets/game-assets/action-icons/strong_defend.png",
    DEBUFF = "/assets/game-assets/action-icons/debuff.png",
    STRONG_DEBUFF = "/assets/game-assets/action-icons/strong_debuff.png",
    BUFF = "/assets/game-assets/action-icons/buff.png",
    STRONG_BUFF = "/assets/game-assets/action-icons/strong_buff.png",
}

export enum backgrounds{
    FIGHT1= "/assets/game-assets/backgrounds/naberoa_tav_v2.png",
    FIGHT2="/assets/game-assets/backgrounds/naberoa_path.png"
}

export enum menuButtons{
    SURRENDER = "/assets/game-assets/gui/surrenderButton.png",
    ATTACK = "/assets/game-assets/gui/attack_button.png",
    GUARD = "/assets/game-assets/gui/guard_button.png",
    ITEM = "/assets/game-assets/gui/item_button.png",
    RUN = "/assets/game-assets/gui/run_button.png",
    MENUBACK = "/assets/game-assets/gui/back_fight_menu.png",
    SOUND= "/assets/game-assets/gui/sound.png",
    SOUNDOFF= "/assets/game-assets/gui/mute.png"
}

export enum playerSprites{
    EXAMPLE = "/assets/game-assets/anacleto_pj.png",
    PLACEHOLDER = "/assets/game-assets/hero-ph.png",
    WARRIOR = "/assets/game-assets/hero-warrior.png",
    MAGE = "/assets/game-assets/hero-mage.png",
    ROGUE = "/assets/game-assets/hero-rogue.png",

}
export enum healthbarEn{
    HEALTH_BAR = "/assets/game-assets/gui/vida.png",
    BAR="/assets/game-assets/gui/healthbar.png"
}

export enum enemyTextures{
    ASSASSIN = "/assets/game-assets/enemy-ph.png",
    SLIME = "/assets/game-assets/slime_ph.png"
}